# api/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Keystone Core REST API
  description: |
    REST API for Keystone Core - a runtime infrastructure control plane.

    This specification covers all HTTP REST endpoints exposed by Keystone Core components.
    The primary control plane operations use gRPC (see proto files), but HTTP REST endpoints
    are available for cluster management, health checks, file operations, and more.

    ## Authentication

    Keystone Core supports multiple authentication methods:
    - **API Key**: Bearer token in `Authorization` header
    - **mTLS**: Client certificate authentication
    - **JWT**: JSON Web Token authentication

    ## Rate Limiting

    API endpoints are subject to rate limiting. When rate limited, the API returns
    HTTP 429 with `Retry-After` header indicating when to retry.

    ## Versioning

    All REST APIs use the `/api/v1/` prefix. Breaking changes will increment the version.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: Keystone Core Team
    url: https://github.com/shawnbutts/keystone-core
servers:
  - url: http://localhost:9090
    description: Local development server
  - url: https://{control-plane-host}:9443
    description: Production server (HTTPS)
    variables:
      control-plane-host:
        default: control-plane.example.com
        description: Control plane hostname
...
security:
  - bearerAuth: []
  - mtls: []
paths:
...
  /api/v1/events:
    get:
      tags:
        - Events
      summary: List events
      description: |
        Returns a list of events with optional filtering.
        Note: Handler exists but is not yet wired into kscore-server.
      operationId: listEvents
      parameters:
        - name: type
          in: query
          description: Filter by event type
          schema:
            type: string
        - name: source
          in: query
          description: Filter by event source
          schema:
            type: string
        - name: since
          in: query
          description: Return events since this time (RFC 3339)
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Events
      summary: Submit event
      description: |
        Submit an event for processing.
        Note: Handler exists but is not yet wired into kscore-server.
      operationId: submitEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '202':
          description: Event accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication
    mtls:
      type: mutualTLS
      description: Mutual TLS client certificate authentication
  parameters:
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 100
        maximum: 1000
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
...
    Event:
      type: object
      required:
        - type
        - source
      properties:
        id:
          type: string
          description: Event ID (generated if not provided)
        type:
          type: string
          description: Event type
        source:
          type: string
          description: Event source
        subject:
          type: string
          description: Event subject
        time:
          type: string
          format: date-time
          description: Event timestamp
        data:
          type: object
          description: Event payload
    EventResponse:
      type: object
      properties:
        id:
          type: string
          description: Assigned event ID
        status:
          type: string
          enum: [accepted, queued]
...